+++ test/widgets/reorderable_builder_item_widget_test.dart
@@ -59,7 +59,7 @@ void main() {
 
   void givenOnDragEnd(
     ReorderableEntity reorderableEntity,
-    Offset globalOffset,
+    Offset? globalOffset,
   ) {}
 
   void givenOnDragCanceled(