+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   - this issue has been resolved to prevent confusion
   - special thanks to `wuhangandroid` and `Gloomy699` ([#57](https://github.com/karvulf/flutter-reorderable-grid-view/issues/57))
 
+⭐️ **New Features**
+- added generic type for `ReorderableBuilder`
+  - that removes the cast of `reorderListFunction` in `onReorder`
+  - special thanks to `KevenMelo` ([#142](https://github.com/karvulf/flutter-reorderable-grid-view/issues/142))
+
 ## 5.3.2
 🐛 **Bug Fixes**
 - **fixed reordering issue**: